requirements.txt
torch>=2.2,<3
torchvision>=0.17,<1
numpy>=1.26,<3
gymnasium>=0.29,<1
minigrid>=2.3,<3
pillow>=10,<12
matplotlib>=3.8,<4
tqdm>=4.66,<5
pyyaml>=6,<7
transformers>=4.44,<5
huggingface-hub>=0.24,<1
safetensors>=0.4,<1
tensorboard>=2.16,<3
wandb>=0.18,<1
